data-processing/2025/202511/20251110/pyproject.toml
[project]
name = "20251103"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.3",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "trino>=0.336.0",
]
